% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\alias{dotplot,data.frame-method}
\alias{dotplot.default}
\title{dotplot for enrichment result}
\usage{
dotplot(obj, ...)

\S4method{dotplot}{data.frame}(obj, ...)

dotplot.default(
  obj,
  x = "omicsRatio",
  color.by = "p.adjust",
  point.size.by = "count.enriched",
  point.size.range = c(3, 8),
  point.color.range = c("red", "blue"),
  order.by = NULL,
  order.decreasing = FALSE,
  order.show = 10,
  font.size = 12,
  x.title = NULL,
  legend.size.title = NULL,
  legend.color.title = NULL,
  legend.position = "right",
  legend.direction = "vertical"
)
}
\arguments{
\item{obj}{a enriched result in data.frame format.}

\item{x}{variable for x-axis, one of 'GeneRatio' (default), 'BgRatio' and 'count.enriched'.}

\item{color.by}{variable that used to color enriched terms,
'arg' should be one of 'pvalue', 'p.adjust' (default), 'qvalue'.}

\item{point.size.by}{variable that used to scale the sizes of categories.}

\item{point.size.range}{a vector of two numbers, set the point size.}

\item{point.color.range}{a vector of two color values, set the colors.}

\item{order.by}{The order of the Y-axis, NULL (default) means no sorting.}

\item{order.decreasing}{logical. Should the order.by order be increasing or decreasing?
This param is invalid when order.by = NULL.}

\item{order.show}{A number or a vector of terms. If it is a number,
the first n terms will be displayed. If it is a vector of terms,
the selected terms will be displayed.}

\item{font.size}{font size}

\item{x.title}{A character string or expression indicating a title of x-axis If NULL, the title is shown by default.}

\item{legend.size.title}{A character string or expression indicating a title of guide (size). If NULL, the title is shown by default.}

\item{legend.color.title}{A character string or expression indicating a title of guide (color). If NULL, the title is shown by default.}

\item{legend.position}{the position of legends ("none", "left", "right" (default), "bottom", "top", or two-element numeric vector \link{0, 1})}

\item{legend.direction}{layout of items in legends, one of "horizontal" or "vertical" (default).}
}
\value{
a \code{ggplot2} instance
}
\description{
plot dotplot for enrichment result by use \code{ggplot2} package.
}
\examples{
# ---------------------------------------------------------------------------
# example one
library(OmicsPEA)

data(Metabolites)
data(Metabolites.Pathways)
colnames(Metabolites.Pathways) <- c('Metabolites', 'itemId', 'pathwayName')
PEA.RES = PEA(omics.id.set = Metabolites$ID, omics.pathway.db = Metabolites.Pathways, min.enrich.size = 1)

dotplot(
  obj = PEA.RES,
  x = 'omicsRatio',
  color.by = 'p.adjust',
  point.size.by = 'count.enriched',
  order.by = 'x',
  order.decreasing = FALSE,
  order.show = c('Phospholipid Metabolism',
                 'Phospholipid Metabolism',
                 'Glycosphingolipid Metabolism',
                 'Vitamin B3 (Niacin, NAD+) Metabolism',
                 'Krebs Cycle',
                 'Histidine, Histamine, Carnosine Metabolism')
)

# ---------------------------------------------------------------------------
# example two
library(OmicsPEA)
library(org.Hs.eg.db)
library(AnnotationDbi)

data(Gene.Expression.Demo)
data(KEGG.Pathways)
colnames(KEGG.Pathways) <- c('itemId', 'gene_symbol', 'PathwayID', 'pathwayName')
Gene.Expression.Demo$gene_id <- mapIds(org.Hs.eg.db, Gene.Expression.Demo$gene_symbol, column = "ENTREZID", keytype = "SYMBOL")
PEA.RES = PEA(omics.id.set = Gene.Expression.Demo$gene_id, omics.pathway.db = KEGG.Pathways, min.enrich.size = 10, ratio.numeric = FALSE)

dotplot(
  obj = PEA.RES,
  x = 'omicsRatio',
  color.by = 'p.adjust',
  point.size.by = 'count.enriched',
  order.by = 'x',
  order.decreasing = TRUE,
  order.show = 10,
  legend.size.title = NULL,
  legend.color.title = NULL,
  legend.position = 'bottom',
  legend.direction = 'horizontal',
) + guides(size  = guide_legend(order = 2), color = guide_colorbar(order = 1)) +
  theme(
   plot.margin = margin(t = 5, r = 20, b = 5, l = 5, unit = 'points')
 )

}
\author{
Abao Xing \url{https://github.com/GitBioinformatics}
}
