% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEA.R
\name{PEA}
\alias{PEA}
\title{Pathway Enrichment Analysis (PEA)}
\usage{
PEA(
  omics.id.set,
  omics.pathway.db,
  adjust.p.method = "BH",
  min.enrich.size = 10,
  max.enrich.size = 1000,
  ratio.numeric = FALSE
)
}
\arguments{
\item{omics.id.set}{required, a vector of gene/protein/metabolite or other omics id.}

\item{omics.pathway.db}{required, It is a pathway database in data.frame format and must contain two columns, \code{itemId} and \code{pathwayName}.}

\item{adjust.p.method}{character, one of "none", "BH (default)", "bonferroni", "fdr", "hochberg", "holm", "hommel", "BY", select method used in p.adjust.}

\item{min.enrich.size}{numeric, minimal size of \code{omics.id.set} annotated by \code{omics.pathway.db}, 10 is default.}

\item{max.enrich.size}{numeric, maximal size of \code{omics.id.set} annotated by \code{omics.pathway.db}, 1000 is default.}

\item{ratio.numeric}{logical, ratio value is in character or numeric format, defaults to FALSE (character).}
}
\value{
a \code{enrich.res} data.frame instance
}
\description{
PEA (Pathway Enrichment Analysis) of a gene/protein/metabolite or other set.
This function will do PEA and return the enriched results.
}
\examples{
# ---------------------------------------------------------------------------
# example one
library(OmicsPEA)

data(Metabolites)
data(Metabolites.Pathways)
colnames(Metabolites.Pathways) <- c('Metabolites', 'itemId', 'pathwayName')
enrich.res = PEA(omics.id.set = Metabolites$ID, omics.pathway.db = Metabolites.Pathways, min.enrich.size = 1)
head(enrich.res)

# ---------------------------------------------------------------------------
# example two
# use gene_symbol as itemId
library(OmicsPEA)

data(Gene.Expression.Demo)
data(Wiki.Pathways)
colnames(Wiki.Pathways) <- c('GeneID', 'itemId', 'PathwayID', 'pathwayName')
enrich.res = PEA(omics.id.set = Gene.Expression.Demo$gene_symbol, omics.pathway.db = Wiki.Pathways, min.enrich.size = 10)
head(enrich.res)

# ---------------------------------------------------------------------------
# example three
# use gene_id as itemId
library(OmicsPEA)
library(org.Hs.eg.db)
library(AnnotationDbi)

data(Gene.Expression.Demo)
data(KEGG.Pathways)
colnames(KEGG.Pathways) <- c('itemId', 'gene_symbol', 'PathwayID', 'pathwayName')
Gene.Expression.Demo$gene_id <- mapIds(org.Hs.eg.db, Gene.Expression.Demo$gene_symbol, column = "ENTREZID", keytype = "SYMBOL")
enrich.res = PEA(omics.id.set = Gene.Expression.Demo$gene_id, omics.pathway.db = KEGG.Pathways, min.enrich.size = 10)
head(enrich.res)

}
\author{
Abao Xing \url{https://github.com/GitBioinformatics}
}
